Separate mesh reading and training objectives
